+++ src/utils/ini.cpp
@@ -10,7 +10,7 @@
 
   */
 
-#include "ini.h"
+#include "utils/ini.h"
 #include "iniDefine.h"
 #include "utils/cIniUtils.h"
